+++ src/main/java/fr/thesmyler/terramap/gui/widgets/markers/markers/AbstractMovingMarkers.java
@@ -13,12 +13,12 @@ public abstract class AbstractMovingMarkers extends Marker {
 	protected float azimuth;
 	protected double oldLongitude, oldLatitude;
 
-	public AbstractMovingMarkers(MarkerController<?> controller, int width, int height, int minZoom, int maxZoom) {
+	public AbstractMovingMarkers(MarkerController<?> controller, float width, float height, int minZoom, int maxZoom) {
 		super(controller, width, height, minZoom, maxZoom);
 		this.movingAnimation = new Animation(10000);
 	}
 
-	public AbstractMovingMarkers(MarkerController<?> controller, int width, int height) {
+	public AbstractMovingMarkers(MarkerController<?> controller, float width, float height) {
 		this(controller, width, height, Integer.MIN_VALUE, Integer.MAX_VALUE);
 	}
